+++ CHANGES.md
@@ -1,3 +1,12 @@
+# Unreleased (plans for v2)
+
+* drop whitelist and only use allowlist
+
+# Unreleased (v1.1.0)
+
+* requires Thredded 1.0.1+
+* use allowlist rather than whitelist and deprecate whitelist
+
 # v1.0.0
 
 Support ruby 3.0